requirements.txt
kafka-python-ng
pyspark==3.5.0
Faker==22.6.0
